fixed an issue after rebase
