+++ models_src/db_inits/tortoise_init.py
@@ -89,6 +89,17 @@ def get_tortoise_config(
         "timezone": "UTC",
     }
 
+from tortoise import Tortoise
+
+
+async def init_db(db_url: str, models: list[str]):
+    await Tortoise.init(db_url=db_url, modules={"models": models})
+    await Tortoise.generate_schemas()
+
+
+async def close_db():
+    await Tortoise.close_connections()
+
 # Async context manager to get a Tortoise connection and register pgvector on demand
 class PgVectorConnection:
     def __init__(self, alias: str = "default"):
